#863 make StringBuilderWriter public and move test
